// src/Upload/dto/createNewEmployee.dto.ts
/* eslint-disable indent */
export class CreateNewEmployee {
  EEID: string;

  FullName: string;

  JobTitle: string;

  Department: string;

  BusinessUnit: string;

  Gender: string;

  Ethnicity: string;

  Age: string;

  HireDate: string;

  AnnualSalary: string;

  Bonus: string;

  Country: string;

  City: string;

  ExitDate: string;

  email: string;
}